Extract the Linking module (account linking) into Api/Linking
Eighth module of the #777 folder migration. Consolidates account-link handling
into Jellyfin.Plugin.SSO_Auth.Api.Linking:

- AccountLinkResolver, AdoptionEligibilityResolver, CanonicalLinkRevoker ->
  Api/Linking/ (namespace Api.Linking), import added at each call site.
- The existing Api/Linking/CanonicalLinkService.cs already lived in the folder
  but still declared the flat Api namespace — a pre-existing folder-vs-namespace
  inconsistency this move corrects (it now declares Api.Linking too, along with
  the CanonicalLinkWriteResult/RemoveResult/Removal types it defines).

Linking may depend on Audit (it audits link writes), Provider (validates the
provider), and RateLimit. Registers those edges on the module DAG rule.

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1578 both TFMs), the module
DAG conformance rule, and the ABI-floor build.

Refs #777
