+++ SSO-Auth.Tests/AccountLinkResolverTests.cs
@@ -1,5 +1,6 @@
 using System;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Linking;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
