+++ SSO-Auth/Api/Shared/FlowResponses.cs
@@ -2,6 +2,7 @@ using System;
 using System.Net.Mime;
 using System.Security.Cryptography;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Linking;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
